refactor: placed all stories on stories folder
