+++ src/stories/Modal.stories.tsx
@@ -1,8 +1,8 @@
 import { Meta, StoryObj } from '@storybook/react';
 
-import Modal from '.';
+import Modal from '~/components/Modal';
 
-import { RequestErrorTemplate } from './templates/RequestError';
+import { RequestErrorTemplate } from '~/components/Modal/templates/RequestError';
 
 import { handleModalRender } from '~/helpers/handleModal';
